requirements.txt
anthropic>=0.39.0
wrds>=3.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
